@@ -1,9 +1,11 @@
+"""Functional utilities."""
+
 from typing import Callable, Sequence, TypeVar
 
 A = TypeVar("A")
 
 
-def take_last_while(predicate: Callable[[A], bool], xs: Sequence[A]) -> Sequence[A]:
+def take_last_while(predicate: Callable[[A], bool], xs: Sequence[A]) -> Sequence[A]:  # noqa: D103
     start = 0
     for e in reversed(xs):
         if not predicate(e):
