@@ -1,3 +1,5 @@
+"""Ids."""
+
 import re
 from typing import Iterable
 
@@ -5,7 +7,7 @@ DIGITS = re.compile(r"\d+")
 
 
 def next_id(xs: Iterable[str]) -> int:
-    """Calculate the next id to use from these identifiers
+    """Calculate the next id to use from these identifiers.
 
     next_id([]) // => 0
     next_id(['t1', 't2', 't3']) // => 4
