@@ -1,10 +1,13 @@
+"""UniqueCheck."""
+
 from typing import Dict, Generic, TypeVar
 
 S = TypeVar("S")
 
 
 class UniqueCheck(Generic[S]):
-    """
+    """Check if values are unique.
+
     >>> u = UniqueCheck()
     >>> u(1)
     True
@@ -20,15 +23,16 @@ class UniqueCheck(Generic[S]):
     False
     """
 
-    def __init__(self) -> None:
+    def __init__(self) -> None:  # noqa: D107
         self.c: Count[S] = Count()
 
-    def __call__(self, s: S) -> bool:
+    def __call__(self, s: S) -> bool:  # noqa: D102
         return self.c.inc(s) == 1
 
 
 class Count(Generic[S]):
-    """
+    """Counter that can be increased and queried.
+
     >>> u = Count()
     >>> u.inc(1)
     1
@@ -50,12 +54,12 @@ class Count(Generic[S]):
     1
     """
 
-    def __init__(self) -> None:
+    def __init__(self) -> None:  # noqa: D107
         self.m: Dict[S, int] = {}
 
-    def get(self, s: S) -> int:
+    def get(self, s: S) -> int:  # noqa: D102
         return self.m.get(s) or 0
 
-    def inc(self, s: S) -> int:
+    def inc(self, s: S) -> int:  # noqa: D102
         self.m[s] = self.get(s) + 1
         return self.get(s)
