@@ -1,7 +1,8 @@
 from typing import List
 
 import pytest
-from parallel_corpus.token import Token, identify, tokenize
+
+from parallel_corpus.text_token import Token, identify, tokenize
 
 
 def test_can_create_token() -> None:
@@ -12,7 +13,7 @@ def test_can_create_token() -> None:
 
 
 @pytest.mark.parametrize(
-    "text, expected",
+    ("text", "expected"),
     [
         ("", []),
         (" ", [" "]),
